// src/generated/types/EventScheduleDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type EventScheduleDto = {
  /**
   * @type string
   */
  labId: string
  /**
   * @type string, date-time
   */
  dateTime: Date
  /**
   * @type string
   */
  examFiles: string
  /**
   * @type string
   */
  assisstantId: string
  /**
   * @type string
   */
  id: string
}